requirements.txt
wheel
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
